Cleaned up a great deal of code by adding comments and also verifying all unit tests pass as expected. Changed all return values to UInt32 in VcuComm code

+++ DLL/VcuComm/CommDevices/ICommDevice.cs
@@ -8,51 +8,52 @@ namespace VcuComm
     /// </summary>
     public interface ICommDevice
     {
+
         /// <summary>
         /// Attempts to opens a port and establish communication with a target device
         /// </summary>
         /// <param name="commaDelimitedOptions">string contains comma delimited options based on the port being opened</param>
         /// <returns>less than 0 if any failure occurs; greater than or equal to 0 if successful</returns>
-        Int16 Open(String commaDelimitedOptions);
+        Int32 Open(String commaDelimitedOptions);
 
         /// <summary>
         /// Sends the start of message delimiter to the target
         /// </summary>
         /// <returns>less than 0 if any failure occurs; greater than or equal to 0 if successful</returns>
-        Int16 SendStartOfMessage();
+        Int32 SendStartOfMessage();
 
         /// <summary>
         /// Received the start of message delimiter from the target. The value of the start of message indicates
         /// the "endianness" of the target.
         /// </summary>
         /// <returns>less than 0 if any failure occurs; greater than or equal to 0 if successful</returns>
-        Int16 ReceiveStartOfMessage();
+        Int32 ReceiveStartOfMessage();
 
         /// <summary>
         /// Sends a message to the target
         /// </summary>
         /// <param name="txMessage">message that is sent to the target</param>
-        /// <returns>less than 0 if any failure occurs; greater than or equal to 0 if successful</returns>
-        Int16 SendDataToTarget(Byte[] txMessage);
+        /// <returns>less than 0 if any failure occurs; number of bytes sent otherwise</returns>
+        Int32 SendDataToTarget(Byte[] txMessage);
 
         /// <summary>
         /// Receives a message from the target
         /// </summary>
         /// <param name="rxMessage">the message received from the target is placed here</param>
         /// <returns>less than 0 if any failure occurs; greater than or equal to 0 if successful</returns>
-        Int16 ReceiveTargetDataPacket(Byte[] rxMessage, out Int32 bytesReceived);
+        Int32 ReceiveTargetDataPacket(Byte[] rxMessage, out Int32 bytesReceived);
 
         /// <summary>
         /// Receives an acknowledge from the target
         /// </summary>
         /// <returns>less than 0 if any failure occurs; greater than or equal to 0 if successful</returns>
-        Int16 ReceiveTargetAcknowledge();
+        Int32 ReceiveTargetAcknowledge();
 
         /// <summary>
         /// Closes the hardware device (TCP connection, RS-232 port, etc.)
         /// </summary>
         /// <returns>less than 0 if any failure occurs; greater than or equal to 0 if successful</returns>
-        Int16 Close();
+        Int32 Close();
 
         /// <summary>
         /// Determines if the target is a big endian or little endian machine
